+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/home/settings/SocialLinkPage.kt
@@ -27,7 +27,6 @@ import com.wafflestudio.snutt2.R
 import com.wafflestudio.snutt2.components.compose.CustomDialog
 import com.wafflestudio.snutt2.components.compose.SimpleTopBar
 import com.wafflestudio.snutt2.lib.android.toast
-import com.wafflestudio.snutt2.lib.facebookLogin
 import com.wafflestudio.snutt2.lib.network.dto.core.UserDto
 import com.wafflestudio.snutt2.ui.SNUTTColors
 import com.wafflestudio.snutt2.ui.SNUTTTypography
